+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.3.0 Add lists by title (2023-01-05)
 
+- added a new tab - singles - for showing books that aren't in series
+- 1.3.1 fix functionality to the desired form, drop link colors
+
 ## 1.2.0 Update to 2022 (2023-01-04)
 
 - authentication with Auth0